Combined Hexatic and Trans into one file
